Move OpenAPI client tests to dedicated test project, cleared up some unrequired internalsVisibleTo usage
