+++ src/JsonApiDotNetCore.OpenApi.Client/AssemblyInfo.cs
@@ -1,0 +1,3 @@
+using System.Runtime.CompilerServices;
+
+[assembly: InternalsVisibleTo("OpenApiClientTests")]
