+++ test/OpenApiClientTests/LegacyClient/GeneratedCode/ILegacyClient.cs
@@ -1,0 +1,9 @@
+using JsonApiDotNetCore.OpenApi.Client;
+
+namespace OpenApiClientTests.LegacyClient.GeneratedCode
+{
+    // ReSharper disable once MemberCanBeInternal
+    partial interface ILegacyClient : IJsonApiClient
+    {
+    }
+}
