--- test/OpenApiTests/ClientLibrary/GeneratedCode/IOpenApiClient.cs
@@ -1,9 +1,0 @@
-using JsonApiDotNetCore.OpenApi.Client;
-
-namespace OpenApiTests.ClientLibrary.GeneratedCode
-{
-    // ReSharper disable once MemberCanBeInternal
-    partial interface IOpenApiClient : IJsonApiClient
-    {
-    }
-}
